requirements.txt
Flask==1.0.2
line-bot-sdk==1.8.0
python-dateutil==2.8.1
Pillow==5.4.1 
pya3rt==1.1  